+++ test/runtests.jl
@@ -12,6 +12,7 @@ using SafeTestsets
 if VERSION >= v"1.6.0"
     @time @safetestset "1.6 Arrows" begin include("newarrows.jl") end
 end
+@time @safetestset "Basic DSL" begin include("dsl.jl") end
 @time @safetestset "Model Construction" begin include("make_model.jl") end
 @time @safetestset "Custom Functions" begin include("custom_functions.jl") end
 @time @safetestset "Model Modification" begin include("model_modification.jl") end