+++ HISTORY.md
@@ -18,6 +18,12 @@
   `Reaction`s work.
 - `ReactionSystem(rxs::Vector{Reaction}, t)` should now work and will infer the
   species and parameters.
+- Added `extend(sys, reactionnetwork, name=nameof(sys))` to extend
+  `ReactionSystem`s with constraint equations (algebraic equations or ODEs).
+  Constraints are stored as a `NonlinearSystem` or `ODESystem` within the
+  `ReactionSystem`, and accessible via `get_constraints(reactionnetwork)`.
+- Added `Catalyst.flatten(rn)` to allow flattening of a `ReactionSystem` with
+  sub-systems into one `ReactionSystem`.
 
 ## Catalyst 9.0
 *1.* **BREAKING:** `netstoichmat`, `prodstoichmat` and `substoichmat` are now